Redistributed ownership of the user defaults key prefix, the Electron now works other than that settings aren't retained.
