+++ OSBindings/Mac/Clock Signal/Machine/CSMachine.h
@@ -33,6 +33,8 @@
 @property (nonatomic, readonly) double clockRate;
 @property (nonatomic, readonly) BOOL clockIsUnlimited;
 
+@property (nonatomic, readonly) NSString *userDefaultsPrefix;
+
 - (void)paste:(NSString *)string;
 
 @end
