+++ OSBindings/Mac/Clock Signal/Machine/StaticAnalyser/CSStaticAnalyser.h
@@ -15,7 +15,6 @@
 - (instancetype)initWithFileAtURL:(NSURL *)url;
 
 @property(nonatomic, readonly) NSString *optionsPanelNibName;
-@property(nonatomic, readonly) NSString *userDefaultsPrefix;
 - (CSMachine *)newMachine;
 
 @property(nonatomic, readonly) NSString *displayName;
