+++ OSBindings/Mac/Clock Signal/Machine/Wrappers/CSVic20.mm
@@ -182,4 +182,6 @@ using namespace Commodore::Vic20;
 	}
 }
 
+- (NSString *)userDefaultsPrefix {	return @"vic20";	}
+
 @end
